+++ src/compiler/SyncerPageCompiler.ts
@@ -16,7 +16,6 @@ import {
 	getRewriteRules,
 	sanitizePermalink,
 } from "../utils/utils";
-import { ExcalidrawCompiler } from "./ExcalidrawCompiler";
 import slugify from "@sindresorhus/slugify";
 import { fixMarkdownHeaderSyntax } from "../utils/markdown";
 import {
@@ -56,7 +55,6 @@ export type TCompilerStep = (
 export class SyncerPageCompiler {
 	private readonly vault: Vault;
 	private readonly settings: QuartzSyncerSettings;
-	private excalidrawCompiler: ExcalidrawCompiler;
 	private metadataCache: MetadataCache;
 	private readonly getFilesMarkedForPublishing: Publisher["getFilesMarkedForPublishing"];
 	private rewriteRule: PathRewriteRule;
@@ -71,7 +69,6 @@ export class SyncerPageCompiler {
 		this.settings = settings;
 		this.metadataCache = metadataCache;
 		this.getFilesMarkedForPublishing = getFilesMarkedForPublishing;
-		this.excalidrawCompiler = new ExcalidrawCompiler(vault);
 		this.rewriteRule = getRewriteRules(this.settings.vaultPath);
 	}
 
@@ -89,18 +86,11 @@ export class SyncerPageCompiler {
 		};
 
 	async generateMarkdown(file: PublishFile): Promise<TCompiledFile> {
-		const assets: Assets = { blobs: [] };
-
 		const vaultFileText = await file.cachedRead();
 
 		if (this.settings.useExcalidraw) {
 			if (file.file.name.endsWith(".excalidraw.md")) {
-				return [
-					await this.excalidrawCompiler.compileMarkdown({
-						includeExcaliDrawJs: true,
-					})(file)(vaultFileText),
-					assets,
-				];
+				console.warn("Excalidraw files are not supported yet.");
 			}
 		}
 
@@ -304,7 +294,6 @@ export class SyncerPageCompiler {
 
 			const transcludedRegex = /!\[\[(.+?)\]\]/g;
 			const transclusionMatches = text.match(transcludedRegex);
-			let numberOfExcaliDraws = 0;
 
 			for (const transclusionMatch of transclusionMatches ?? []) {
 				try {
@@ -347,22 +336,7 @@ export class SyncerPageCompiler {
 					});
 
 					if (linkedFile.name.endsWith(".excalidraw.md")) {
-						numberOfExcaliDraws++;
-						const isFirstDrawing = numberOfExcaliDraws === 1;
-
-						const fileText = await publishLinkedFile.cachedRead();
-
-						const excaliDrawCode =
-							await this.excalidrawCompiler.compileMarkdown({
-								includeExcaliDrawJs: isFirstDrawing,
-								idAppendage: `${numberOfExcaliDraws}`,
-								includeFrontMatter: false,
-							})(publishLinkedFile)(fileText);
-
-						transcludedText = transcludedText.replace(
-							transclusionMatch,
-							excaliDrawCode,
-						);
+						continue;
 					} else if (linkedFile.extension === "md") {
 						let fileText = await publishLinkedFile.cachedRead();