+++ docs/Changelog.md
@@ -2,7 +2,7 @@
 title: Roadmap
 description: Changelog and feature roadmap for Quartz Syncer.
 created: 2025-05-16T12:59:31Z+0200
-modified: 2025-05-26T14:28:53Z+0200
+modified: 2025-05-26T15:12:43Z+0200
 publish: true
 ---
 
@@ -25,6 +25,7 @@ publish: true
 ### Version 1.6.4
 
 - Moved all inlined styling to CSS classes.
+- Removed broken Excalidraw stub.
 
 ### Version 1.6.3